--HG--
branch : binary_serialization

+++ src/ompl/contrib/rrt_star/src/RRTstar.cpp
@@ -72,7 +72,7 @@ void ompl::geometric::RRTstar::setup(void)
     sc.configurePlannerRange(maxDistance_);
 
     ballRadiusMax_ = si_->getMaximumExtent();
-    ballRadiusConst_ = maxDistance_ * sqrt(si_->getStateSpace()->getDimension());
+    ballRadiusConst_ = maxDistance_ * sqrt((double)si_->getStateSpace()->getDimension());
 
     delayCC_ = true;
